Big Refactoring: deleted Repository with implementation via JdbcTemplate , left Repository with implementation via Hibernate. Renamed classes, moved to new directory
