+++ myBrikProject/src/main/java/by/mybrik/repository/GoodsRepository.java
@@ -1,6 +1,7 @@
 package by.mybrik.repository;
 
-import by.mybrik.domain.Goods;
+import by.mybrik.domain.entities.Goods;
 
-public interface GoodsRepository extends CrudRepository <Long, Goods>{
+public interface GoodsRepository extends CrudRepository<Long, Goods> {
+
 }
