--- myBrikProject/src/main/java/by/mybrik/repository/TextileProductTypeRepository.java
@@ -1,6 +1,0 @@
-package by.mybrik.repository;
-
-import by.mybrik.domain.TextileProductType;
-
-public interface TextileProductTypeRepository extends CrudRepository <Long, TextileProductType> {
-}
