+++ myBrikProject/src/main/java/by/mybrik/repository/TextileRepository.java
@@ -1,6 +1,6 @@
 package by.mybrik.repository;
 
-import by.mybrik.domain.Textile;
+import by.mybrik.domain.entities.Textile;
 
-public interface TextileRepository extends CrudRepository <Long, Textile>{
+public interface TextileRepository extends CrudRepository<Long, Textile> {
 }
