+++ myBrikProject/src/main/java/by/mybrik/repository/implementation/GoodsImpl.java
@@ -1,8 +1,7 @@
-package by.mybrik.repository.newImplementation.implementation;
+package by.mybrik.repository.implementation;
 
-import by.mybrik.domain.entities.Textile;
-import by.mybrik.repository.newImplementation.TextileRep;
-import lombok.RequiredArgsConstructor;
+import by.mybrik.domain.entities.Goods;
+import by.mybrik.repository.GoodsRepository;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
@@ -14,58 +13,61 @@ import java.util.Optional;
 
 @Repository
 @Primary
-@RequiredArgsConstructor
-public class TextileImpl implements TextileRep {
+public class GoodsImpl implements GoodsRepository {
 
   private final SessionFactory sessionFactory;
 
+  public GoodsImpl(SessionFactory sessionFactory) {
+    this.sessionFactory = sessionFactory;
+  }
+
   @Override
-  public Textile save(Textile textile) {
-    try (Session session = sessionFactory.openSession()) {
-      session.saveOrUpdate(textile);
-      return textile;
+  public Goods save(Goods product) {
+    try (final Session session = sessionFactory.openSession()) {
+      session.saveOrUpdate(product);
+      return product;
     }
   }
 
   @Override
-  public List<Textile> findAll() {
+  public List<Goods> findAll() {
     try (Session session = sessionFactory.openSession()) {
-      String hqlQuery = "select u from Textile u";
-      return session.createQuery(hqlQuery, Textile.class).list();
+      String hqlQuery = "select u from Goods u";
+      return session.createQuery(hqlQuery, Goods.class).list();
     }
   }
 
   @Override
-  public Textile findById(Long key) {
+  public Goods findById(Long key) {
     try (Session session = sessionFactory.openSession()) {
-      return session.find(Textile.class, key);
+      return session.find(Goods.class, key);
     }
   }
 
   @Override
-  public Optional<Textile> findOne(Long key) {
+  public Optional<Goods> findOne(Long key) {
     return Optional.empty();
   }
 
   @Override
-  public Textile update(Textile textile) {
+  public Goods update(Goods product) {
     try (Session session = sessionFactory.openSession()) {
       Transaction transaction = session.getTransaction();
       transaction.begin();
-      session.saveOrUpdate(textile);
+      session.saveOrUpdate(product);
       transaction.commit();
-      return textile;
+      return product;
     }
   }
 
   @Override
-  public Long delete(Textile textile) {
+  public Long delete(Goods product) {
     try (Session session = sessionFactory.openSession()) {
       Transaction transaction = session.getTransaction();
       transaction.begin();
-      session.delete(textile);
+      session.delete(product);
       transaction.commit();
-      return textile.getId();
+      return product.getId();
     }
   }
 }
