+++ myBrikProject/src/main/java/by/mybrik/repository/implementation/ProductTypeImpl.java
@@ -1,8 +1,7 @@
-package by.mybrik.repository.newImplementation.implementation;
+package by.mybrik.repository.implementation;
 
-import by.mybrik.domain.entities.StandardOrder;
-import by.mybrik.repository.newImplementation.StandardOrderRep;
-import lombok.RequiredArgsConstructor;
+import by.mybrik.domain.entities.ProductType;
+import by.mybrik.repository.ProductTypeRepository;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
@@ -14,58 +13,61 @@ import java.util.Optional;
 
 @Repository
 @Primary
-@RequiredArgsConstructor
-public class StandardOrderImpl implements StandardOrderRep {
+public class ProductTypeImpl implements ProductTypeRepository {
 
   private final SessionFactory sessionFactory;
 
+  public ProductTypeImpl(SessionFactory sessionFactory) {
+    this.sessionFactory = sessionFactory;
+  }
+
   @Override
-  public StandardOrder save(StandardOrder order) {
+  public ProductType save(ProductType type) {
     try (Session session = sessionFactory.openSession()) {
-      session.saveOrUpdate(order);
-      return order;
+      session.saveOrUpdate(type);
+      return type;
     }
   }
 
   @Override
-  public List<StandardOrder> findAll() {
+  public List<ProductType> findAll() {
     try (Session session = sessionFactory.openSession()) {
-      String hqlQuery = "select u from StandardOrder u";
-      return session.createQuery(hqlQuery, StandardOrder.class).list();
+      String hqlQuery = "select u from ProductType u";
+      return session.createQuery(hqlQuery, ProductType.class).list();
     }
   }
 
   @Override
-  public StandardOrder findById(Long key) {
+  public ProductType findById(Long key) {
     try (Session session = sessionFactory.openSession()) {
-      return session.find(StandardOrder.class, key);
+      return session.find(ProductType.class, key);
     }
   }
 
   @Override
-  public Optional<StandardOrder> findOne(Long key) {
+  public Optional<ProductType> findOne(Long key) {
     return Optional.empty();
   }
 
   @Override
-  public StandardOrder update(StandardOrder order) {
+  public ProductType update(ProductType type) {
     try (Session session = sessionFactory.openSession()) {
       Transaction transaction = session.getTransaction();
       transaction.begin();
-      session.saveOrUpdate(order);
+      session.saveOrUpdate(type);
       transaction.commit();
-      return order;
+      return type;
     }
   }
 
   @Override
-  public Long delete(StandardOrder order) {
+  public Long delete(ProductType type) {
     try (Session session = sessionFactory.openSession()) {
       Transaction transaction = session.getTransaction();
       transaction.begin();
-      session.delete(order);
+      session.delete(type);
       transaction.commit();
-      return order.getId();
+      return type.getId();
     }
   }
 }
