--- myBrikProject/src/main/java/by/mybrik/repository/newImplementation/GoodsRep.java
@@ -1,8 +1,0 @@
-package by.mybrik.repository.newImplementation;
-
-import by.mybrik.domain.entities.Goods;
-import by.mybrik.repository.CrudRepository;
-
-public interface GoodsRep extends CrudRepository<Long, Goods> {
-
-}
