--- myBrikProject/src/main/java/by/mybrik/repository/newImplementation/IndividualOrderRep.java
@@ -1,7 +1,0 @@
-package by.mybrik.repository.newImplementation;
-
-import by.mybrik.domain.entities.IndividualOrder;
-import by.mybrik.repository.CrudRepository;
-
-public interface IndividualOrderRep extends CrudRepository<Long, IndividualOrder> {
-}
