--- myBrikProject/src/main/java/by/mybrik/repository/newImplementation/PriceForIndividualOrderRep.java
@@ -1,7 +1,0 @@
-package by.mybrik.repository.newImplementation;
-
-import by.mybrik.domain.entities.PriceForIndividualOrder;
-import by.mybrik.repository.CrudRepository;
-
-public interface PriceForIndividualOrderRep extends CrudRepository<Long, PriceForIndividualOrder> {
-}
