--- myBrikProject/src/main/java/by/mybrik/repository/newImplementation/ProductTypeRep.java
@@ -1,7 +1,0 @@
-package by.mybrik.repository.newImplementation;
-
-import by.mybrik.domain.entities.ProductType;
-import by.mybrik.repository.CrudRepository;
-
-public interface ProductTypeRep extends CrudRepository<Long, ProductType> {
-}
