--- myBrikProject/src/main/java/by/mybrik/repository/newImplementation/StandardOrderRep.java
@@ -1,7 +1,0 @@
-package by.mybrik.repository.newImplementation;
-
-import by.mybrik.domain.entities.StandardOrder;
-import by.mybrik.repository.CrudRepository;
-
-public interface StandardOrderRep extends CrudRepository<Long, StandardOrder> {
-}
