--- myBrikProject/src/main/java/by/mybrik/repository/newImplementation/TextileRep.java
@@ -1,7 +1,0 @@
-package by.mybrik.repository.newImplementation;
-
-import by.mybrik.domain.entities.Textile;
-import by.mybrik.repository.CrudRepository;
-
-public interface TextileRep extends CrudRepository<Long, Textile> {
-}
