--- myBrikProject/src/main/java/by/mybrik/repository/newImplementation/UsersRep.java
@@ -1,7 +1,0 @@
-package by.mybrik.repository.newImplementation;
-
-import by.mybrik.domain.entities.Users;
-import by.mybrik.repository.CrudRepository;
-
-public interface UsersRep extends CrudRepository<Long, Users> {
-}
