+++ myBrikProject/src/main/java/by/mybrik/service/impl/StandardOrderServiceImpl.java
@@ -1,7 +1,7 @@
 package by.mybrik.service.impl;
 
 import by.mybrik.domain.entities.StandardOrder;
-import by.mybrik.repository.newImplementation.StandardOrderRep;
+import by.mybrik.repository.StandardOrderRepository;
 import by.mybrik.service.StandardOrderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,7 +13,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class StandardOrderServiceImpl implements StandardOrderService {
 
-  private final StandardOrderRep standardOrderRepository;
+  private final StandardOrderRepository standardOrderRepository;
 
   @Override
   public StandardOrder save(StandardOrder order) {
